components: rename
