+++ src/components/app/app.tsx
@@ -3,8 +3,8 @@ import {Route, Routes, BrowserRouter} from 'react-router-dom';
 import MainPage from '../../pages/main/main-page';
 import {AppRoute} from '../../consts';
 import NotFoundPage from '../../pages/not-found/not-found';
-import LoginPage from '../../pages/login/login-page';
-import FavoritesPage from '../../pages/favorites/favorites-page';
+import LoginPage from '../../pages/login-page/login-page';
+import FavoritesPage from '../../pages/favorites-page/favorites-page';
 import OfferPage from '../../pages/offer/offer-page';
 import PrivateRoute from '../private-route/private-route';
 import {useSelector} from 'react-redux';
